ao/openal: Added option to set buffers characteristics.
One can now set the number of buffers, the buffer size, and the period
size.
This can reduce the CPU usage and the total latency stays mostly the same.
As there are sync mechanisms the A/V sync continue intact and working.

It also modifies 6.1 channel order, as per OpenAL spec
and add AOPLAY_FINAL_CHUNK support