[ao/openal] Added option to set buffers characteristics.
One can now set the number of buffers, the buffer size, and the period
size. Also, as OpenAL does not have a limit on the buffer size all
available samples are buffered in one go.
This can reduce the CPU usage and the total latency stays mostly the same.
As there are sync mechanisms the A/V sync continue intact and working.